Crate rome_diagnostics

Re-exports

pub use file::Span;

pub use termcolor;

Modules

file

v2

Structs

CodeSuggestion

A Suggestion that is provided by rslint, and can be reported to the user, and can be automatically applied if it has the right Applicability.

Diagnostic

A diagnostic message that can give information like errors or warnings.

Emitter

The emitter is responsible for emitting diagnostics to a given output.

Footer

A note or help that is displayed under the diagnostic.

SubDiagnostic

Everything that can be added to a diagnostic, like a suggestion that will be displayed under the actual error.

Enums

Applicability

Indicates how a tool should manage this suggestion.

DiagnosticTag

Severity

A severity level for diagnostic messages.

Constants

MAXIMUM_DISPLAYABLE_DIAGNOSTICS