Trait rome_diagnostics::file::Files

pub trait Files {
    fn name(&self, id: FileId) -> Option<&str>;
    fn source(&self, id: FileId) -> Option<SourceFile<'_>>;
}

Interface for interacting with source files that are identified by a unique identifier.

Required Methods

fn name(&self, id: FileId) -> Option<&str>

Returns the name of the file identified by the id.

fn source(&self, id: FileId) -> Option<SourceFile<'_>>

Returns the source of the file identified by the id.