Trait rome_diagnostics::file::Span

pub trait Span {
    fn as_range(&self) -> TextRange;

    fn join<T: Span>(&self, other: T) -> TextRange { ... }
    fn between<T: Span>(&self, other: T) -> TextRange { ... }
    fn until<T: Span>(&self, other: T) -> TextRange { ... }
    fn sub_start(&self, amount: TextSize) -> TextRange { ... }
    fn add_start(&self, amount: TextSize) -> TextRange { ... }
    fn sub_end(&self, amount: TextSize) -> TextRange { ... }
    fn add_end(&self, amount: TextSize) -> TextRange { ... }
}

A value which can be used as the range inside of a diagnostic.

This is essentially a hack to allow us to use SyntaxElement, SyntaxNode, etc directly

Required Methods

fn as_range(&self) -> TextRange

Provided Methods

fn join<T: Span>(&self, other: T) -> TextRange

Make a new span which extends to another span

from      to
^^^^^^^^^^^^

fn between<T: Span>(&self, other: T) -> TextRange

Make a new span which is between another span

from      to
    ^^^^^^

fn until<T: Span>(&self, other: T) -> TextRange

Make a new span which extends until another span

from      to
^^^^^^^^^^

fn sub_start(&self, amount: TextSize) -> TextRange

fn add_start(&self, amount: TextSize) -> TextRange

fn sub_end(&self, amount: TextSize) -> TextRange

fn add_end(&self, amount: TextSize) -> TextRange