Struct rome_diagnostics::Diagnostic

pub struct Diagnostic {
    pub file_id: FileId,
    pub severity: Severity,
    pub code: Option<&'static Category>,
    pub title: MarkupBuf,
    pub summary: Option<String>,
    pub tag: Option<DiagnosticTag>,
    pub primary: Option<SubDiagnostic>,
    pub children: Vec<SubDiagnostic>,
    pub suggestions: Vec<CodeSuggestion>,
    pub footers: Vec<Footer>,
}

A diagnostic message that can give information like errors or warnings.

Fields
file_id: FileIdseverity: Severitycode: Option<&'static Category>title: MarkupBufsummary: Option<String>tag: Option<DiagnosticTag>primary: Option<SubDiagnostic>children: Vec<SubDiagnostic>suggestions: Vec<CodeSuggestion>footers: Vec<Footer>
Implementations

impl Diagnostic

pub fn error(
file_id: FileId,
code: &'static Category,
title: impl Display
) -> Self

Creates a new Diagnostic with the Error severity.

pub fn warning(
file_id: FileId,
code: &'static Category,
title: impl Display
) -> Self

Creates a new Diagnostic with the Warning severity.

pub fn help(file_id: FileId, code: &'static Category, title: impl Display) -> Self

Creates a new Diagnostic with the Help severity.

pub fn note(file_id: FileId, code: &'static Category, title: impl Display) -> Self

Creates a new Diagnostic with the Note severity.

pub fn bug(file_id: FileId, code: &'static Category, title: impl Display) -> Self

Creates a new Diagnostic with the Bug severity.

pub fn new(file_id: FileId, severity: Severity, title: impl Display) -> Self

Creates a new Diagnostic that will be used in a builder-like way to modify labels, and suggestions.

pub fn new_with_code(
file_id: FileId,
severity: Severity,
title: impl Display,
code: Option<&'static Category>
) -> Self

Creates a new Diagnostic with an error code that will be used in a builder-like way to modify labels, and suggestions.

pub fn severity(self, severity: Severity) -> Self

Overwrites the severity of this diagnostic.

pub fn summary(self, summary: impl Into<String>) -> Self

Set an explicit plain-text summary for this diagnostic.

pub fn deprecated(self) -> Self

Marks this diagnostic as deprecated code, which will be displayed in the language server.

This does not have any influence on the diagnostic rendering.

pub fn unnecessary(self) -> Self

Marks this diagnostic as unnecessary code, which will be displayed in the language server.

This does not have any influence on the diagnostic rendering.

pub fn label_in_file(
self,
severity: Severity,
span: FileSpan,
msg: impl Display
) -> Self

Attaches a label to this Diagnostic, that will point to another file that is provided.

pub fn label(self, severity: Severity, span: impl Span, msg: impl Display) -> Self

Attaches a label to this Diagnostic.

The given span has to be in the file that was provided while creating this Diagnostic.

pub fn primary(self, span: impl Span, msg: impl Display) -> Self

Attaches a primary label to this Diagnostic.

pub fn secondary(self, span: impl Span, msg: impl Display) -> Self

Attaches a secondary label to this Diagnostic.

pub fn suggestion_full(
self,
span: impl Span,
msg: impl Display,
suggestion: TextEdit,
applicability: Applicability
) -> Self

Add a suggestion to this Diagnostic

pub fn footer(self, severity: Severity, msg: impl Display) -> Self

Adds a footer to this Diagnostic, which will be displayed under the actual error.

pub fn footer_help(self, msg: impl Display) -> Self

Adds a footer to this Diagnostic, with the Help severity.

pub fn footer_note(self, msg: impl Display) -> Self

Adds a footer to this Diagnostic, with the Note severity.

pub fn is_error(&self) -> bool

Checks if the severity of the current diagnostic is Severity::Error or higher

impl Diagnostic

pub fn as_diagnostic<'a>(&'a self, files: &'a dyn Files) -> impl Diagnostic + 'a

pub fn display<'a>(&'a self, files: &'a dyn Files) -> impl Display + 'a