Trait rome_diagnostics::v2::advice::Visit

pub trait Visit {
    fn record_log(
        &mut self,
        category: LogCategory,
        text: &dyn Display
    ) -> Result<()> { ... }
    fn record_list(&mut self, list: &[&dyn Display]) -> Result<()> { ... }
    fn record_frame(&mut self, location: Location<'_>) -> Result<()> { ... }
    fn record_diff(&mut self, diff: &TextEdit) -> Result<()> { ... }
    fn record_backtrace(
        &mut self,
        title: &dyn Display,
        backtrace: &Backtrace
    ) -> Result<()> { ... }
    fn record_command(&mut self, command: &str) -> Result<()> { ... }
    fn record_group(
        &mut self,
        title: &dyn Display,
        advice: &dyn Advices
    ) -> Result<()> { ... }
}

The Visit trait is used to collect advices from a diagnostic: a visitor instance is provided to the Diagnostic::advices and Diagnostic::verbose_advices methods, and the diagnostic implementation is expected to call into the various record_* methods to communicate advices to the user.

Provided Methods

fn record_log(&mut self, category: LogCategory, text: &dyn Display) -> Result<()>

Prints a single log entry with the provided category and markup.

fn record_list(&mut self, list: &[&dyn Display]) -> Result<()>

Prints an unordered list of items.

fn record_frame(&mut self, location: Location<'_>) -> Result<()>

Prints a code frame outlining the provided source location.

fn record_diff(&mut self, diff: &TextEdit) -> Result<()>

Prints the diff between the prev and next strings.

fn record_backtrace(
&mut self,
title: &dyn Display,
backtrace: &Backtrace
) -> Result<()>

Prints a Rust backtrace.

fn record_command(&mut self, command: &str) -> Result<()>

Prints a command to the user.

fn record_group(
&mut self,
title: &dyn Display,
advice: &dyn Advices
) -> Result<()>

Prints a group of advices under a common title.