Trait rome_diagnostics::v2::context::Context

pub trait Context<T, E>: Sealed {
    fn context<M>(self, message: M) -> Result<T, Error>
    where
        E: 'static,
        M: Display + 'static,
        Error: From<ContextDiagnostic<M, E>>;
    fn with_category(self, category: &'static Category) -> Result<T, Error>
    where
        Error: From<CategoryDiagnostic<E>>;
    fn with_file_path(self, path: impl AsResource) -> Result<T, Error>
    where
        Error: From<FilePathDiagnostic<E>>;
}

Required Methods

fn context<M>(self, message: M) -> Result<T, Error> where
E: 'static,
M: Display + 'static,
Error: From<ContextDiagnostic<M, E>>,

If self is an error, returns a new diagnostic with the provided message as a message and description, and self as a source diagnostic. This is useful to create chains of diagnostics, where high level errors wrap lower level causes.

fn with_category(self, category: &'static Category) -> Result<T, Error> where
Error: From<CategoryDiagnostic<E>>,

If self is an error, returns a new diagnostic using the provided category if self doesn’t already have one.

fn with_file_path(self, path: impl AsResource) -> Result<T, Error> where
Error: From<FilePathDiagnostic<E>>,

If self is an error, returns a new diagnostic using the provided path if self doesn’t already have one.