Trait rome_diagnostics::v2::diagnostic::Diagnostic

pub trait Diagnostic: Debug {
    fn category(&self) -> Option<&'static Category> { ... }
    fn severity(&self) -> Severity { ... }
    fn description(&self, fmt: &mut Formatter<'_>) -> Result { ... }
    fn message(&self, fmt: &mut Formatter<'_>) -> Result<()> { ... }
    fn advices(&self, visitor: &mut dyn Visit) -> Result<()> { ... }
    fn verbose_advices(&self, visitor: &mut dyn Visit) -> Result<()> { ... }
    fn location(&self) -> Option<Location<'_>> { ... }
    fn tags(&self) -> DiagnosticTags { ... }
    fn source(&self) -> Option<&dyn Diagnostic> { ... }
}

The Diagnostic trait defines the metadata that can be exposed by error types in order to print details diagnostics in the console of the editor

Implementation

Most types should not have to implement this trait manually, and should instead rely on the Diagnostic derive macro also provided by this crate:

#[derive(Debug, Diagnostic)]
#[diagnostic(category = "lint/style/noShoutyConstants", tags(FIXABLE))]
struct ExampleDiagnostic {
    #[message]
    #[description]
    message: String,
}

Provided Methods

fn category(&self) -> Option<&'static Category>

The category of a diagnostic uniquely identifying this diagnostic type, such as lint/correctness/noArguments, args/invalid or format/disabled.

fn severity(&self) -> Severity

The severity defines whether this diagnostic reports an error, a warning, an information or a hint to the user.

fn description(&self, fmt: &mut Formatter<'_>) -> Result

The description is a text-only explanation of the issue this diagnostic is reporting, intended for display contexts that do not support rich markup such as in-editor popovers

The description should generally be as exhaustive as possible, since the clients that do not support rendering markup will not render the advices for the diagnostic either.

fn message(&self, fmt: &mut Formatter<'_>) -> Result<()>

An explanation of the issue this diagnostic is reporting

In general it’s better to keep this message as short as possible, and instead rely on advices to better convey contextual explanations to the user.

fn advices(&self, visitor: &mut dyn Visit) -> Result<()>

Advices are the main building blocks used compose rich errors. They are implemented using a visitor pattern, where consumers of a diagnostic can visit the object and collect the advices that make it up for the purpose of display or introspection.

fn verbose_advices(&self, visitor: &mut dyn Visit) -> Result<()>

Diagnostics can defines additional advices to be printed if the user requires more detail about the diagnostic.

fn location(&self) -> Option<Location<'_>>

A diagnostic can be tied to a specific “location”: this can be a file, memory buffer, command line argument, etc. It may also be tied to a specific text range within the content of that location. Finally, it may also provide the source string for that location (this is required in order to display a code frame advice for the diagnostic).

fn tags(&self) -> DiagnosticTags

Tags convey additional boolean metadata about the nature of a diagnostic:

If the diagnostic can be automatically fixed

If the diagnostic resulted from and internal error

If the diagnostic is being emitted as part of a crash / fatal error

If the diagnostic is a warning about a piece of unused or unnecessary code

If the diagnostic is a warning about a piece of deprecated or obsolete code.

fn source(&self) -> Option<&dyn Diagnostic>

Similarly to the source method of the std::error::Error trait, this returns another diagnostic that’s the logical “cause” for this issue. For instance, a “request failed” diagnostic may have been cause by a “deserialization error”. This allows low-level error to be wrapped in higher level concepts, while retaining enough information to display and fix the underlying issue.