Module rome_diagnostics::v2::error

The error module contains the implementation of Error, a dynamic container struct for any type implementing Diagnostic.

The Error struct is essentially a manual implementation of Box<dyn Diagnostic> that is only a single word in size (thin pointer) instead of two for Rust’s dyn types (wide pointer). This is done by manually managing how the “vtable” (the function pointer table used for dynamic dispatch) for the type is laid out and accessed in memory, at the cost of requiring the use of unsafe code.

While reducing the size of Error is the main reason this type is using unsafe code (as it makes returning a Result<T, Error> more efficient), manually managing the vtable for Error opens additional possibilities for future extensions to the type like requiring disjoint trait bounds or implementing dynamic dispatch to traits that require handling owned instances of an object (the main use case for this would be implementing Clone for Error since dyn Clone is not allowed in Rust)

Structs

Error

The Error struct wraps any type implementing Diagnostic into a single dynamic type.

Type Definitions

Result

Alias of std::result::Result with the Err type defaulting to Error.