Struct rome_diagnostics::v2::error::Error

pub struct Error { /* private fields */ }

The Error struct wraps any type implementing Diagnostic into a single dynamic type.

Implementations

impl Error

Implement the Diagnostic trait as inherent methods on the Error type.

pub fn category(&self) -> Option<&Category>

Calls Diagnostic::category on the Diagnostic wrapped by this Error.

pub fn severity(&self) -> Severity

Calls Diagnostic::severity on the Diagnostic wrapped by this Error.

pub fn description(&self, fmt: &mut Formatter<'_>) -> Result

Calls Diagnostic::description on the Diagnostic wrapped by this Error.

pub fn message(&self, fmt: &mut Formatter<'_>) -> Result<()>

Calls Diagnostic::message on the Diagnostic wrapped by this Error.

pub fn advices(&self, visitor: &mut dyn Visit) -> Result<()>

Calls Diagnostic::advices on the Diagnostic wrapped by this Error.

pub fn verbose_advices(&self, visitor: &mut dyn Visit) -> Result<()>

Calls Diagnostic::verbose_advices on the Diagnostic wrapped by this Error.

pub fn location(&self) -> Option<Location<'_>>

Calls Diagnostic::location on the Diagnostic wrapped by this Error.

pub fn tags(&self) -> DiagnosticTags

Calls Diagnostic::tags on the Diagnostic wrapped by this Error.

pub fn source(&self) -> Option<&dyn Diagnostic>

Calls Diagnostic::source on the Diagnostic wrapped by this Error.