Module rome_diagnostics::v2::location

Structs

FileId

An id that points into a file database.

LineIndex

LineIndexBuf

Location

Represents the location of a diagnostic in a resource.

LocationBuilder

Builder type for the Location struct

SourceCode

Represents the source code of a file.

Enums

FilePath

Represents the path of a file on the filesystem.

Resource

Represents the resource a diagnostic is associated with.

Traits

AsResource

Utility trait for types that can be converted to a Resource

AsSourceCode

Utility trait for types that can be converted into SourceCode

AsSpan

Utility trait for types that can be converted into Option<TextRange>