Module rome_diagnostics::v2::prelude

Anonymously re-exports all the traits declared by this module, this is intended to be imported as use rome_diagnostics::v2::prelude::*; to automatically bring all these traits into the ambient context

Re-exports

pub use super::advice::Advices as _;

pub use super::advice::Visit as _;

pub use super::context::Context as _;

pub use super::context::DiagnosticExt as _;

pub use super::diagnostic::Diagnostic as _;